delete gating service
