--- embrace-android-core/src/main/kotlin/io/embrace/android/embracesdk/internal/gating/Sanitizable.kt
@@ -1,6 +1,0 @@
-package io.embrace.android.embracesdk.internal.gating
-
-internal interface Sanitizable<T> {
-
-    fun sanitize(): T?
-}
