+++ embrace-android-core/src/main/kotlin/io/embrace/android/embracesdk/internal/session/message/PayloadMessageCollatorImpl.kt
@@ -1,7 +1,6 @@
 package io.embrace.android.embracesdk.internal.session.message
 
 import io.embrace.android.embracesdk.internal.envelope.session.SessionEnvelopeSource
-import io.embrace.android.embracesdk.internal.gating.GatingService
 import io.embrace.android.embracesdk.internal.payload.Envelope
 import io.embrace.android.embracesdk.internal.payload.SessionPayload
 import io.embrace.android.embracesdk.internal.prefs.PreferencesService
@@ -12,7 +11,6 @@ import io.embrace.android.embracesdk.internal.spans.CurrentSessionSpan
  * Generates a payload
  */
 internal class PayloadMessageCollatorImpl(
-    private val gatingService: GatingService,
     private val sessionEnvelopeSource: SessionEnvelopeSource,
     private val preferencesService: PreferencesService,
     private val currentSessionSpan: CurrentSessionSpan,
@@ -31,13 +29,10 @@ internal class PayloadMessageCollatorImpl(
     }
 
     override fun buildFinalEnvelope(params: FinalEnvelopeParams): Envelope<SessionPayload> {
-        val envelope = gatingService.gateSessionEnvelope(
-            hasCrash = params.crashId != null,
-            envelope = sessionEnvelopeSource.getEnvelope(
-                endType = params.endType,
-                startNewSession = params.startNewSession,
-                crashId = params.crashId
-            )
+        val envelope = sessionEnvelopeSource.getEnvelope(
+            endType = params.endType,
+            startNewSession = params.startNewSession,
+            crashId = params.crashId
         )
         return Envelope<SessionPayload>(
             // future work: make legacy fields null here.
