+++ embrace-android-payload/src/main/kotlin/io/embrace/android/embracesdk/internal/config/instrumented/schema/InstrumentedConfig.kt
@@ -11,6 +11,5 @@ interface InstrumentedConfig {
     val otelLimits: OtelLimitsConfig
     val project: ProjectConfig
     val redaction: RedactionConfig
-    val session: SessionConfig
     val symbols: Base64SharedObjectFilesMap
 }
