+++ embrace-android-payload/src/main/kotlin/io/embrace/android/embracesdk/internal/config/remote/SessionRemoteConfig.kt
@@ -12,19 +12,4 @@ import com.squareup.moshi.JsonClass
 data class SessionRemoteConfig(
     @Json(name = "enable")
     val isEnabled: Boolean? = null,
-
-    /**
-     * A list of session components (i.e. Breadcrumbs, Session properties, etc) that will be
-     * included in the session payload. If components list exists, the services should restrict
-     * the data that is provided to the session.
-     */
-    @Json(name = "components")
-    val sessionComponents: Set<String>? = null,
-
-    /**
-     * A list of session components allowed to send a full session payload (only if "components"
-     * exists)
-     */
-    @Json(name = "send_full_for")
-    val fullSessionEvents: Set<String>? = null,
 )
