+++ embrace-gradle-plugin-integration-tests/src/test/java/io/embrace/android/gradle/integration/testcases/ConfigInstrumentationTest.kt
@@ -26,7 +26,6 @@ class ConfigInstrumentationTest {
         "OtelLimitsConfigImpl",
         "ProjectConfigImpl",
         "RedactionConfigImpl",
-        "SessionConfigImpl",
     ).map {
         "/io/embrace/android/embracesdk/internal/config/instrumented/$it"
     }
