+++ embrace-test-fakes/src/main/kotlin/io/embrace/android/embracesdk/fakes/Behavior.kt
@@ -44,7 +44,7 @@ fun createAnrBehavior(
  */
 fun createSessionBehavior(
     remoteCfg: RemoteConfig? = null,
-): SessionBehavior = SessionBehaviorImpl(InstrumentedConfigImpl, remoteCfg)
+): SessionBehavior = SessionBehaviorImpl(remoteCfg)
 
 /**
  * A [NetworkBehaviorImpl] that returns default values.
