+++ embrace-test-fakes/src/main/kotlin/io/embrace/android/embracesdk/fakes/behavior/FakeSessionBehavior.kt
@@ -1,7 +1,6 @@
 package io.embrace.android.embracesdk.fakes.behavior
 
 import io.embrace.android.embracesdk.internal.config.behavior.SessionBehavior
-import io.embrace.android.embracesdk.internal.config.instrumented.schema.SessionConfig
 import io.embrace.android.embracesdk.internal.config.remote.RemoteConfig
 
 class FakeSessionBehavior(
@@ -9,20 +8,10 @@ class FakeSessionBehavior(
     private val sessionControlEnabled: Boolean = false,
 ) : SessionBehavior {
 
-    override val local: SessionConfig
-        get() = throw UnsupportedOperationException()
+    override val local: Unit = Unit
     override val remote: RemoteConfig
         get() = throw UnsupportedOperationException()
 
-    override fun getFullSessionEvents(): Set<String> = emptySet()
-    override fun getSessionComponents(): Set<String>? = null
-    override fun isGatingFeatureEnabled(): Boolean = false
     override fun isSessionControlEnabled(): Boolean = sessionControlEnabled
     override fun getMaxSessionProperties(): Int = maxSessionProperties
-    override fun shouldGateInfoLog(): Boolean = false
-    override fun shouldGateWarnLog(): Boolean = false
-    override fun shouldSendFullForCrash(): Boolean = false
-    override fun shouldSendFullForErrorLog(): Boolean = false
-    override fun shouldGateSessionProperties(): Boolean = false
-    override fun shouldGateLogProperties(): Boolean = false
 }
