+++ embrace-test-fakes/src/main/kotlin/io/embrace/android/embracesdk/fakes/config/FakeInstrumentedConfig.kt
@@ -7,7 +7,6 @@ import io.embrace.android.embracesdk.internal.config.instrumented.schema.Network
 import io.embrace.android.embracesdk.internal.config.instrumented.schema.OtelLimitsConfig
 import io.embrace.android.embracesdk.internal.config.instrumented.schema.ProjectConfig
 import io.embrace.android.embracesdk.internal.config.instrumented.schema.RedactionConfig
-import io.embrace.android.embracesdk.internal.config.instrumented.schema.SessionConfig
 
 /**
  * A fake [InstrumentedConfig] implementation that defaults to the real implementation unless an override is supplied.
@@ -20,7 +19,6 @@ data class FakeInstrumentedConfig(
     override val otelLimits: OtelLimitsConfig = FakeOtelLimitsConfig(),
     override val project: ProjectConfig = FakeProjectConfig(base.project, appId = "abcde"),
     override val redaction: RedactionConfig = FakeRedactionConfig(base.redaction),
-    override val session: SessionConfig = FakeSessionConfig(base.session),
     override val symbols: Base64SharedObjectFilesMap =
         FakeBase64SharedObjectFilesMap(base.symbols.getBase64SharedObjectFilesMap()),
 ) : InstrumentedConfig
